fix: create event card timeline
